requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
Pillow==10.0.1
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
resend